requirements.txt
apiclient==1.0.4
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cmdstanpy==1.2.0
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.0
google-api-core==2.15.0
google-api-python-client==1.7.3
google-auth>=1.5.0,<2dev
google-auth-httplib2==0.0.3
google-auth-oauthlib>=0.2.0
googleapis-common-protos==1.62.0
holidays==0.40
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
patsy==0.5.5
pillow==10.2.0
prophet==1.1.5
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycausalimpact==0.1.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.4
six==1.16.0
stanio==0.3.0
statsmodels==0.14.1
tqdm==4.66.1
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
